+++ src/Surfnet/StepupRa/RaBundle/Service/SecondFactorService.php
@@ -51,11 +51,15 @@ class SecondFactorService
 
     /**
      * @param string $registrationCode
+     * @param string $actorInstitution
+     * @param string $actorId
      * @return null|VerifiedSecondFactor
      */
-    public function findVerifiedSecondFactorByRegistrationCode($registrationCode)
+    public function findVerifiedSecondFactorByRegistrationCode($registrationCode, $actorInstitution, $actorId)
     {
         $query = new VerifiedSecondFactorSearchQuery();
+        $query->setActorInstitution($actorInstitution);
+        $query->setActorId($actorId);
         $query->setRegistrationCode($registrationCode);
 
         try {